+++ docs/source/changelog.rst
@@ -5,7 +5,8 @@ Change Log
 **Pyro 4.50**
 
 - Future can now be given a delay before it is evaluated
-
+- new ITER_STREAM_LINGER config item to keep streams alive for a given period after proxy disconnect (default=0, no lingering)
+  @todo docs
 
 **Pyro 4.49**